Refactor youtube service to be more in same code pattern as other services
